Enable SPI4 interface, add to mcu_init.
The SPI4 that is used for the MPU9250 on the Lisa/S V2.0 was not
enabled. It was just leading to an obscure system crash within chibios.

The SPI4 interface was implemented in the lowlevel driver but the init
function was never called from the top level mcu_init function...